+++ bench/src/main/scala/me/shadaj/scalapy/py/DataSendingBench.scala
@@ -1,6 +1,5 @@
 package me.shadaj.scalapy.py
 
-import PyConverters._
 import me.shadaj.scalapy.interpreter.PyValue
 
 object DataSendingBench extends App {